New changes

+++ src/main/java/com/example/baygo/db/service/impl/OrderServiceImpl.java
@@ -2,6 +2,8 @@ package com.example.baygo.db.service.impl;
 
 import com.example.baygo.db.config.jwt.JwtService;
 import com.example.baygo.db.custom.CustomOrderRepository;
+import com.example.baygo.db.exceptions.BadCredentialException;
+import com.example.baygo.db.exceptions.NotFoundException;
 import com.example.baygo.db.model.Seller;
 import com.example.baygo.db.model.User;
 import com.example.baygo.db.model.enums.Status;
@@ -36,19 +38,33 @@ public class OrderServiceImpl implements OrderService {
     @Override
     public SimpleResponse deleteById(Long orderId) {
         User user = jwtService.getAuthenticate();
+        if (user == null) {
+            return SimpleResponse.builder()
+                    .message("Authentication required")
+                    .status(HttpStatus.UNAUTHORIZED)
+                    .build();
+        }
+
         Seller seller = user.getSeller();
+        if (seller == null) {
+            return SimpleResponse.builder()
+                    .message("Seller not found")
+                    .status(HttpStatus.NOT_FOUND)
+                    .build();
+        }
+
         try {
             orderRepository.deleteById(orderId);
             return SimpleResponse.builder()
                     .message(String.format("Order %s deleted successfully", orderId))
                     .status(HttpStatus.OK)
                     .build();
-        } catch (EmptyResultDataAccessException e) {
+        } catch (NotFoundException e) {
             return SimpleResponse.builder()
                     .message(String.format("Order with ID %s not found", orderId))
                     .status(HttpStatus.NOT_FOUND)
                     .build();
-        } catch (Exception e) {
+        } catch (BadCredentialException e) {
             return SimpleResponse.builder()
                     .message("Failed to delete order")
                     .status(HttpStatus.INTERNAL_SERVER_ERROR)
@@ -57,8 +73,10 @@ public class OrderServiceImpl implements OrderService {
     }
 
     @Override
-    public AnalysisResponse getWeeklyAnalisys(Date startDate, Date endDate, Long warehouseId, String nameOfTime,boolean commission) {
-          return customOrderRepository.getWeeklyAnalysis(startDate,endDate,warehouseId,nameOfTime,commission);
+    public AnalysisResponse getWeeklyAnalisys(Date startDate, Date endDate, Long warehouseId, String nameOfTime) {
+        User user = jwtService.getAuthenticate();
+        Seller seller = user.getSeller();
+          return customOrderRepository.getWeeklyAnalysis(startDate,endDate,warehouseId,nameOfTime, seller.getId());
     }
